Refactored hero-detail into own vue component
